fix: limit Lexical internal links to pages and posts
Timeslot admin overrides were replacing plugin defaults and leaving timeslots linkable, while simpleLexical used an unrestricted LinkFeature — so internal doc pickers defaulted to timeslots.
